Extract field-utils from repo
